Refactor code
- Add some methods.
- Replace several methods.
